# Don't evaluate regexes twice