--- a/build.gradle
+++ b/build.gradle
@@ -47,7 +47,7 @@ subprojects {
     sourceCompatibility = 1.8
 
     tasks.withType(JavaCompile) {
-        options.compilerArgs += ['-Xlint:deprecation', '-Werror']
+        options.compilerArgs += ['-Werror']
     }
 
     plugins.withId('com.palantir.baseline-error-prone', {