--- a/changelog/@unreleased/pr-462.v2.yml
+++ b/changelog/@unreleased/pr-462.v2.yml
@@ -1,0 +1,6 @@
+type: improvement
+improvement:
+  description: Parsing an `OrderableSlsVersion` no longer runs the regex matcher twice,
+    reducing allocation slightly
+  links:
+  - https://github.com/palantir/sls-version-java/pull/462