--- a/sls-versions/src/main/java/com/palantir/sls/versions/NonOrderableSlsVersion.java
+++ b/sls-versions/src/main/java/com/palantir/sls/versions/NonOrderableSlsVersion.java
@@ -22,6 +22,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.palantir.logsafe.UnsafeArg;
 import java.util.Optional;
+import java.util.regex.Matcher;
 import org.immutables.value.Value;
 
 @Value.Immutable
@@ -41,16 +42,16 @@ public abstract class NonOrderableSlsVersion extends SlsVersion {
             return Optional.empty();
         }
 
-        RegexGroups groups = SlsVersionType.NON_ORDERABLE.getPattern().tryParse(value);
+        Matcher groups = SlsVersionType.NON_ORDERABLE.getPattern().tryParse(value);
         if (groups == null) {
             return Optional.empty();
         }
 
         return Optional.of(new NonOrderableSlsVersion.Builder()
                 .value(value)
-                .majorVersionNumber(groups.groupAsInt(1))
-                .minorVersionNumber(groups.groupAsInt(2))
-                .patchVersionNumber(groups.groupAsInt(3))
+                .majorVersionNumber(Integer.parseInt(groups.group(1)))
+                .minorVersionNumber(Integer.parseInt(groups.group(2)))
+                .patchVersionNumber(Integer.parseInt(groups.group(3)))
                 .type(SlsVersionType.NON_ORDERABLE)
                 .build());
     }