--- a/sls-versions/src/main/java/com/palantir/sls/versions/SlsVersionMatcherParser.java
+++ b/sls-versions/src/main/java/com/palantir/sls/versions/SlsVersionMatcherParser.java
@@ -121,7 +121,7 @@ final class SlsVersionMatcherParser {
             try {
                 return ok(next, Integer.parseUnsignedInt(string.substring(startIndex, next)));
             } catch (NumberFormatException e) {
-                if (e.getMessage().endsWith("exceeds range of unsigned int.")) {
+                if (e.getMessage() != null && e.getMessage().endsWith("exceeds range of unsigned int.")) {
                     return fail(startIndex);
                 } else {
                     throw e;