--- a/versions.lock
+++ b/versions.lock
@@ -1,5 +1,6 @@
 # Run ./gradlew --write-locks to regenerate this file
 com.fasterxml.jackson.core:jackson-annotations:2.11.1 (2 constraints: bb17a070)
+com.google.code.findbugs:jsr305:3.0.2 (1 constraints: 170aecb4)
 com.google.errorprone:error_prone_annotations:2.5.1 (2 constraints: 1b1bf559)
 com.palantir.safe-logging:preconditions:1.4.0 (1 constraints: 0705fc35)
 com.palantir.safe-logging:safe-logging:1.14.0 (2 constraints: 3816b8f5)
@@ -13,7 +14,6 @@ org.slf4j:slf4j-api:1.7.30 (1 constraints: 3d05453b)
 [Test dependencies]
 com.fasterxml.jackson.core:jackson-core:2.11.1 (1 constraints: 85123021)
 com.fasterxml.jackson.core:jackson-databind:2.11.1 (1 constraints: 3705333b)
-com.google.code.findbugs:jsr305:3.0.2 (1 constraints: 170aecb4)
 com.google.guava:failureaccess:1.0.1 (1 constraints: 140ae1b4)
 com.google.guava:guava:30.1.1-jre (1 constraints: a5064e53)
 com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava (1 constraints: bd17c918)