--- a/versions.props
+++ b/versions.props
@@ -4,6 +4,7 @@ org.slf4j:slf4j-api = 1.7.30
 com.palantir.safe-logging:preconditions = 1.4.0
 com.palantir.safe-logging:safe-logging = 1.14.0
 com.google.guava:guava = 30.1.1-jre
+com.uber.nullaway:nullaway = 0.8.0
 
 # Test dependencies
 org.assertj:assertj-core = 3.19.0