## Before this PR

The GC problems on the service I work on are a lot better, but not completely solved. There are some sensible domain-related changes we could make, but something that stuck out to me when looking at the JFRs was how much costly it was to just parse version numbers (in terms of CPU time and also allocation)

![image](https://user-images.githubusercontent.com/3473798/121976969-d51cb800-cd7c-11eb-96be-7d5cb79465ff.png)

Interestingly we've had this comment `Note: Can avoid duplicate regex evaluate if performance ever becomes a concern.`

## After this PR
==COMMIT_MSG==
Parsing an `OrderableSlsVersion` no longer runs the regex matcher twice
==COMMIT_MSG==

```
BEFORE


Benchmark                                                         (versionString)  Mode  Cnt     Score      Error   Units
SlsVersionBenchmark.safeValueOf                                           RELEASE  avgt    3   278.398 ±    1.942   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                            RELEASE  avgt    3  7413.325 ±   80.262  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                       RELEASE  avgt    3   632.032 ±    0.011    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                   RELEASE  avgt    3  7448.247 ±  240.664  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm              RELEASE  avgt    3   635.010 ±   15.795    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space               RELEASE  avgt    3     0.033 ±    0.043  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm          RELEASE  avgt    3     0.003 ±    0.004    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                 RELEASE  avgt    3   177.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                                  RELEASE  avgt    3    97.000                 ms
SlsVersionBenchmark.safeValueOf                                          SNAPSHOT  avgt    3   909.686 ±  177.282   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                           SNAPSHOT  avgt    3  4797.870 ±  932.620  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                      SNAPSHOT  avgt    3  1336.090 ±    0.040    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                  SNAPSHOT  avgt    3  4821.985 ±   58.242  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm             SNAPSHOT  avgt    3  1342.908 ±  263.204    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space              SNAPSHOT  avgt    3     0.023 ±    0.048  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm         SNAPSHOT  avgt    3     0.006 ±    0.012    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                SNAPSHOT  avgt    3   138.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                                 SNAPSHOT  avgt    3    75.000                 ms
SlsVersionBenchmark.safeValueOf                                                RC  avgt    3   534.555 ±  135.397   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                                 RC  avgt    3  5473.985 ± 1400.198  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                            RC  avgt    3   896.061 ±    0.016    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                        RC  avgt    3  5520.608 ± 1141.179  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm                   RC  avgt    3   903.744 ±  168.509    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space                    RC  avgt    3     0.031 ±    0.114  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm               RC  avgt    3     0.005 ±    0.019    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                      RC  avgt    3   158.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                                       RC  avgt    3    85.000                 ms
SlsVersionBenchmark.safeValueOf                                       RC_SNAPSHOT  avgt    3   678.839 ±  170.689   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                        RC_SNAPSHOT  avgt    3  4197.744 ± 1011.516  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                   RC_SNAPSHOT  avgt    3   872.060 ±    0.017    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space               RC_SNAPSHOT  avgt    3  4232.266 ± 1097.996  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm          RC_SNAPSHOT  avgt    3   879.280 ±  219.905    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space           RC_SNAPSHOT  avgt    3     0.020 ±    0.042  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm      RC_SNAPSHOT  avgt    3     0.004 ±    0.008    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                             RC_SNAPSHOT  avgt    3   121.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                              RC_SNAPSHOT  avgt    3    65.000                 ms
SlsVersionBenchmark.safeValueOf                                             DIRTY  avgt    3   632.662 ±  956.822   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                              DIRTY  avgt    3  4522.205 ± 6689.555  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                         DIRTY  avgt    3   872.060 ±    0.009    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                     DIRTY  avgt    3  4545.028 ± 7768.420  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm                DIRTY  avgt    3   875.854 ±  209.024    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space                 DIRTY  avgt    3     0.019 ±    0.056  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm            DIRTY  avgt    3     0.004 ±    0.005    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                   DIRTY  avgt    3   130.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                                    DIRTY  avgt    3    73.000                 ms

Process finished with exit code 0

AFTER


Benchmark                                                         (versionString)  Mode  Cnt     Score       Error   Units
SlsVersionBenchmark.safeValueOf                                           RELEASE  avgt    3   194.011 ±   501.334   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                            RELEASE  avgt    3  7094.521 ± 17288.513  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                       RELEASE  avgt    3   416.019 ±     0.007    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                   RELEASE  avgt    3  7120.312 ± 16316.958  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm              RELEASE  avgt    3   417.846 ±    82.823    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space               RELEASE  avgt    3     0.024 ±     0.136  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm          RELEASE  avgt    3     0.001 ±     0.005    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                 RELEASE  avgt    3   140.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                                  RELEASE  avgt    3    93.000                  ms
SlsVersionBenchmark.safeValueOf                                          SNAPSHOT  avgt    3   714.961 ±   402.728   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                           SNAPSHOT  avgt    3  5119.310 ±  2900.577  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                      SNAPSHOT  avgt    3  1120.075 ±     0.029    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                  SNAPSHOT  avgt    3  5139.262 ±  1851.504  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm             SNAPSHOT  avgt    3  1124.726 ±   287.525    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space              SNAPSHOT  avgt    3     0.023 ±     0.073  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm         SNAPSHOT  avgt    3     0.005 ±     0.013    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                SNAPSHOT  avgt    3   147.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                                 SNAPSHOT  avgt    3    81.000                  ms
SlsVersionBenchmark.safeValueOf                                                RC  avgt    3   443.385 ±   577.418   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                                 RC  avgt    3  5024.689 ±  6824.128  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                            RC  avgt    3   680.038 ±     0.004    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                        RC  avgt    3  5057.595 ±  7982.883  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm                   RC  avgt    3   684.097 ±   145.258    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space                    RC  avgt    3     0.016 ±     0.110  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm               RC  avgt    3     0.002 ±     0.012    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                      RC  avgt    3   120.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                                       RC  avgt    3    77.000                  ms
SlsVersionBenchmark.safeValueOf                                       RC_SNAPSHOT  avgt    3   711.226 ±   410.655   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                        RC_SNAPSHOT  avgt    3  4010.460 ±  2271.405  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                   RC_SNAPSHOT  avgt    3   872.060 ±     0.036    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space               RC_SNAPSHOT  avgt    3  4026.378 ±  2207.875  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm          RC_SNAPSHOT  avgt    3   875.543 ±    60.157    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space           RC_SNAPSHOT  avgt    3     0.016 ±     0.072  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm      RC_SNAPSHOT  avgt    3     0.003 ±     0.014    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                             RC_SNAPSHOT  avgt    3   115.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                              RC_SNAPSHOT  avgt    3    66.000                  ms
SlsVersionBenchmark.safeValueOf                                             DIRTY  avgt    3   616.208 ±   740.972   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                              DIRTY  avgt    3  4634.911 ±  5512.971  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                         DIRTY  avgt    3   872.055 ±     0.111    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                     DIRTY  avgt    3  4645.809 ±  5495.985  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm                DIRTY  avgt    3   874.135 ±    88.477    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space                 DIRTY  avgt    3     0.019 ±     0.087  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm            DIRTY  avgt    3     0.004 ±     0.014    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                   DIRTY  avgt    3   122.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                                    DIRTY  avgt    3    72.000                  ms

```

## Possible downsides?
- this might be excessive... my normal rule of thumb is to avoid microoptimising unless you can get at least an order of magnitude