@carterkozak 2021-06-15
lgtm pending splitting the parse method